feat: refactor config and crate dependencies
