+++ crates/aggchain-proof-builder/Cargo.toml
@@ -10,13 +10,11 @@ anyhow.workspace = true
 futures.workspace = true
 thiserror.workspace = true
 tower = { workspace = true, features = ["timeout"] }
-tracing.workspace = true
 serde = { workspace = true, features = ["derive"] }
-sp1-sdk = { workspace = true, features = ["native-gnark"] }
+sp1-sdk.workspace = true
 url.workspace = true
 
 aggchain-proof-core.workspace = true
-aggkit-prover-config.workspace = true
 
 [lints]
 workspace = true
