+++ crates/aggchain-proof-service/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "aggchain-proof-service"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+serde.workspace = true
+thiserror.workspace = true
+tower = { workspace = true, features = ["timeout"] }
+tracing.workspace = true
+
+aggchain-proof-builder.workspace = true
+aggkit-prover-types.workspace = true
+proposer-service.workspace = true
+
+[features]
+default = []
+testutils = []
