+++ crates/aggchain-proof-service/src/error.rs
