+++ crates/aggchain-proof-service/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod config;
+
+pub mod error;
+
+pub mod service;
