+++ crates/aggkit-prover-config/Cargo.toml
@@ -4,19 +4,11 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
-dirs.workspace = true
-ethers.workspace = true
-humantime-serde = "1.1.1"
-jsonrpsee.workspace = true
 serde = { workspace = true, features = ["derive"] }
-serde_with.workspace = true
-sp1-sdk = { workspace = true, features = ["native-gnark"] }
 thiserror.workspace = true
 toml.workspace = true
-tracing.workspace = true
-tracing-subscriber = { workspace = true, features = ["env-filter"] }
-url = { workspace = true, features = ["serde"] }
 
+aggchain-proof-service.workspace = true
 prover-logger.workspace = true
 prover-utils.workspace = true
 prover-config.workspace = true
@@ -24,9 +16,4 @@ prover-config.workspace = true
 [dev-dependencies]
 insta = { workspace = true, features = ["toml", "redactions", "filters"] }
 pretty_assertions = "1.4.0"
-rstest.workspace = true
-serde_json = { workspace = true }
 
-[features]
-default = []
-testutils = []
