--- crates/aggkit-prover/src/aggchain_proof/mod.rs
@@ -1,8 +1,0 @@
-pub mod error;
-pub(crate) mod rpc;
-pub(crate) mod service;
-
-pub use rpc::GrpcService;
-
-#[cfg(test)]
-mod tests;
