+++ crates/aggkit-prover/src/main.rs
@@ -1,6 +1,5 @@
 use std::sync::Arc;
 
-use aggkit_prover::aggchain_proof::GrpcService;
 use aggkit_prover_types::v1::aggchain_proof_service_server::AggchainProofServiceServer;
 use prover_engine::ProverEngine;
 use tokio_util::sync::CancellationToken;
@@ -28,8 +27,9 @@ fn main() -> anyhow::Result<()> {
         .build()?;
 
     // TODO: implement the aggchain-proof service
-    let aggchain_proof_service =
-        AggchainProofServiceServer::new(GrpcService::new(&config.aggchain_proof_service)?);
+    let aggchain_proof_service = AggchainProofServiceServer::new(
+        aggkit_prover::rpc::GrpcService::new(&config.aggchain_proof_service)?,
+    );
 
     _ = ProverEngine::builder()
         .add_rpc_service(aggchain_proof_service)
