+++ crates/agglayer-prover-config/Cargo.toml
@@ -4,17 +4,9 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
-dirs.workspace = true
-ethers.workspace = true
-humantime-serde = "1.1.1"
-jsonrpsee.workspace = true
 serde = { workspace = true, features = ["derive"] }
-serde_with.workspace = true
 thiserror.workspace = true
 toml.workspace = true
-tracing.workspace = true
-tracing-subscriber = { workspace = true, features = ["env-filter"] }
-url = { workspace = true, features = ["serde"] }
 
 prover-logger.workspace = true
 prover-utils.workspace = true
@@ -23,9 +15,4 @@ prover-config.workspace = true
 [dev-dependencies]
 insta = { workspace = true, features = ["toml", "redactions", "filters"] }
 pretty_assertions = "1.4.0"
-rstest.workspace = true
-serde_json = { workspace = true }
 
-[features]
-default = []
-testutils = []
