+++ crates/agglayer-prover/Cargo.toml
@@ -11,12 +11,9 @@ workspace = true
 anyhow.workspace = true
 bincode.workspace = true
 buildstructor.workspace = true
-futures.workspace = true
-thiserror.workspace = true
+sp1-sdk.workspace = true
 tokio = { workspace = true, features = ["full"] }
 tokio-util.workspace = true
-toml.workspace = true
-tracing-subscriber = { workspace = true, features = ["env-filter", "json"] }
 tracing.workspace = true
 tower = { workspace = true, features = ["timeout"] }
 tonic = { workspace = true, features = ["zstd"] }
@@ -30,8 +27,7 @@ prover-engine.workspace = true
 prover-executor.workspace = true
 prover-logger.workspace = true
 
-sp1-sdk = { workspace = true, features = ["native-gnark"] }
-sp1-prover = { workspace = true, features = ["native-gnark"] }
+
 
 [features]
 default = []
