+++ crates/proposer-client/Cargo.toml
@@ -11,17 +11,12 @@ workspace = true
 anyhow.workspace = true
 alloy-primitives.workspace = true
 hex.workspace = true
-aggkit-prover-config.workspace = true
 reqwest = { workspace = true, features = ["json"]}
 serde.workspace = true
+sp1-sdk.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 tonic.workspace = true
 
 
-sp1-sdk = { workspace = true, features = ["native-gnark"] }
 
-
-[features]
-default = []
-testutils = []
